requirements.txt
streamlit
ultralytics
opencv-python-headless
matplotlib
pillow
numpy
